list command works now
hi this took several hours but it works now, only bad thing is that if a username is all numbers (like 25), it doesn't work xD, am working on a fix but this should be pushed since it works for 99.9% of users. Am really happy to have a fully functional list command now that even generates an image for the thing! :3